requirements.txt
loguru==0.6.0
PyExecJS==1.5.1
requests==2.28.1
